+++ bonded-roles/src/main/java/bisq/bonded_roles/explorer/ExplorerService.java
@@ -21,132 +21,42 @@ import bisq.bonded_roles.explorer.dto.Tx;
 import bisq.common.data.Pair;
 import bisq.common.json.JsonMapperProvider;
 import bisq.common.threading.ExecutorFactory;
-import bisq.common.util.ExceptionUtil;
-import bisq.i18n.Res;
-import bisq.network.BaseService;
+import bisq.network.HttpRequestBaseService;
 import bisq.network.NetworkService;
 import bisq.network.http.BaseHttpClient;
-import bisq.network.http.utils.HttpException;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionException;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.RejectedExecutionException;
-import java.util.concurrent.atomic.AtomicInteger;
-import static com.google.common.base.Preconditions.checkNotNull;
-import static java.util.concurrent.TimeUnit.SECONDS;
+
 
 // TODO We should support same registration model via oracle node as used with other nodes
 
 @Slf4j
-public class ExplorerService extends BaseService {
+public class ExplorerService extends HttpRequestBaseService {
     private static final ExecutorService POOL = ExecutorFactory.newCachedThreadPool("ExplorerService", 1, 5, 60);
 
 
     public ExplorerService(Config conf, NetworkService networkService) {
-        super("api/tx/", conf, networkService);
-    }
-
-    @Override
-    public CompletableFuture<Boolean> shutdown() {
-        shutdownStarted = true;
-        return httpClient.map(BaseHttpClient::shutdown)
-                .orElse(CompletableFuture.completedFuture(true));
+        super("api/tx/", conf, networkService, POOL);
     }
 
     public CompletableFuture<Tx> requestTx(String txId) {
-        try {
-            return requestTx(txId, new AtomicInteger(0))
-                    .exceptionallyCompose(throwable -> {
-                        if (throwable instanceof RetryException retryException) {
-                            return requestTx(txId, retryException.getRecursionDepth());
-                        } else if (ExceptionUtil.getRootCause(throwable) instanceof RetryException retryException) {
-                            return requestTx(txId, retryException.getRecursionDepth());
-                        } else {
-                            return CompletableFuture.failedFuture(throwable);
-                        }
-                    });
-        } catch (RejectedExecutionException e) {
-            return CompletableFuture.failedFuture(new RejectedExecutionException("Too many requests. Try again later."));
-        }
-    }
-
-    public String getSelectedProviderBaseUrl() {
-        return Optional.ofNullable(selectedProvider.get()).map(ExplorerService.Provider::getBaseUrl).orElse(Res.get("data.na"));
-    }
-
-    private CompletableFuture<Tx> requestTx(String txId, AtomicInteger recursionDepth) {
-        if (noProviderAvailable) {
-            return CompletableFuture.failedFuture(new RuntimeException("No block explorer provider available"));
-        }
-        if (shutdownStarted) {
-            return CompletableFuture.failedFuture(new RuntimeException("Shutdown has already started"));
-        }
-
-        try {
-            return CompletableFuture.supplyAsync(() -> {
-                        Provider provider = checkNotNull(selectedProvider.get(), "Selected provider must not be null.");
-                        BaseHttpClient client = networkService.getHttpClient(provider.getBaseUrl(), userAgent, provider.getTransportType());
-                        httpClient = Optional.of(client);
-                        long ts = System.currentTimeMillis();
-                        String param = provider.getApiPath() + txId;
-                        try {
-                            log.info("Request tx with ID {} from {}", txId, client.getBaseUrl() + "/" + param);
-                            String json = client.get(param, Optional.of(new Pair<>("User-Agent", userAgent)));
-                            log.info("Received tx lookup response from {}/{} after {} ms", client.getBaseUrl(), param, System.currentTimeMillis() - ts);
-                            selectedProvider.set(selectNextProvider());
-                            shutdownHttpClient(client);
-                            return JsonMapperProvider.get().readValue(json, Tx.class);
-                        } catch (Exception e) {
-                            shutdownHttpClient(client);
-                            if (shutdownStarted) {
-                                throw new RuntimeException("Shutdown has already started");
-                            }
-
-                            Throwable rootCause = ExceptionUtil.getRootCause(e);
-                            log.warn("Encountered exception for TxId={} from provider {}", txId, provider.getBaseUrl(), rootCause);
-
-                            if (rootCause instanceof HttpException httpException) {
-                                int responseCode = httpException.getResponseCode();
-                                // If not server error we pass the error to the client
-                                // 408 (Request Timeout) and 429 (Too Many Requests) are usually transient
-                                // and should rotate to another provider.
-                                if (responseCode < 500 && responseCode != 408 && responseCode != 429) {
-                                    throw new CompletionException(e);
-                                }
-                            }
-
-                            int numRecursions = recursionDepth.incrementAndGet();
-                            if (numRecursions < numTotalCandidates && failedProviders.size() < numTotalCandidates) {
-                                failedProviders.add(provider);
-                                selectedProvider.set(selectNextProvider());
-                                log.warn("We retry the request with new provider {}", selectedProvider.get().getBaseUrl());
-                                throw new RetryException("Retrying with next provider", recursionDepth);
-                            } else {
-                                log.warn("We exhausted all possible providers and give up");
-                                throw new RuntimeException("We failed at all possible providers and give up");
-                            }
-                        }
-                    }, POOL)
-                    .completeOnTimeout(null, conf.getTimeoutInSeconds(), SECONDS)
-                    .thenCompose(tx -> {
-                        if (tx == null) {
-                            return CompletableFuture.failedFuture(new RetryException("Timeout", recursionDepth));
-                        }
-                        return CompletableFuture.completedFuture(tx);
-                    });
-        } catch (RejectedExecutionException e) {
-            log.error("Executor rejected requestTx task. TxId={}", txId, e);
-            return CompletableFuture.failedFuture(e);
-        }
-    }
-
-    private void shutdownHttpClient(BaseHttpClient client) {
-        try {
-            client.shutdown();
-        } catch (Exception ignore) {
-        }
+        return executeWithRetry(() -> makeRequest(provider -> {
+            long ts = System.currentTimeMillis();
+            String param = provider.getApiPath() + txId;
+            log.info("Request tx with ID {} from {}", txId, provider.getBaseUrl() + "/" + param);
+            
+            BaseHttpClient client = networkService.getHttpClient(provider.getBaseUrl(), userAgent, provider.getTransportType());
+            try {
+                String json = client.get(param, Optional.of(new Pair<>("User-Agent", userAgent)));
+                log.info("Received tx lookup response from {}/{} after {} ms", provider.getBaseUrl(), param, System.currentTimeMillis() - ts);
+                
+                return CompletableFuture.completedFuture(JsonMapperProvider.get().readValue(json, Tx.class));
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }));
     }
 }
