+++ network/network/src/main/java/bisq/network/http/utils/ReferenceTimeService.java
@@ -20,9 +20,7 @@ package bisq.network.http.utils;
 import bisq.common.data.Pair;
 import bisq.common.json.JsonMapperProvider;
 import bisq.common.threading.ExecutorFactory;
-import bisq.common.util.ExceptionUtil;
-import bisq.i18n.Res;
-import bisq.network.BaseService;
+import bisq.network.HttpRequestBaseService;
 import bisq.network.NetworkService;
 import bisq.network.http.BaseHttpClient;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -31,127 +29,39 @@ import lombok.extern.slf4j.Slf4j;
 import java.util.Date;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionException;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.RejectedExecutionException;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-import static java.util.concurrent.TimeUnit.SECONDS;
 
 /**
  * Derived from ExplorerService
  */
 @Slf4j
-public class ReferenceTimeService extends BaseService {
+public class ReferenceTimeService extends HttpRequestBaseService {
     private static final ExecutorService POOL = ExecutorFactory.newCachedThreadPool("ReferenceTimeService", 1, 4, 60);
 
     public ReferenceTimeService(Config conf, NetworkService networkService) {
-       super("api/v1/prices", conf, networkService);
-    }
-
-    @Override
-    public CompletableFuture<Boolean> shutdown() {
-        shutdownStarted = true;
-        return httpClient.map(BaseHttpClient::shutdown)
-                .orElse(CompletableFuture.completedFuture(true));
+       super("api/v1/prices", conf, networkService, POOL);
     }
 
     public CompletableFuture<Long> request() {
-        try {
-            return request(new AtomicInteger(0))
-                    .exceptionallyCompose(throwable -> {
-                        if (throwable instanceof RetryException retryException) {
-                            return request(retryException.getRecursionDepth());
-                        } else if (ExceptionUtil.getRootCause(throwable) instanceof RetryException retryException) {
-                            return request(retryException.getRecursionDepth());
-                        } else {
-                            return CompletableFuture.failedFuture(throwable);
-                        }
-                    });
-        } catch (RejectedExecutionException e) {
-            return CompletableFuture.failedFuture(new RejectedExecutionException("Too many requests. Try again later."));
-        }
-    }
-
-    public String getSelectedProviderBaseUrl() {
-        return Optional.ofNullable(selectedProvider.get()).map(ReferenceTimeService.Provider::getBaseUrl).orElse(Res.get("data.na"));
-    }
-
-    private CompletableFuture<Long> request(AtomicInteger recursionDepth) {
-        if (noProviderAvailable) {
-            return CompletableFuture.failedFuture(new RuntimeException("No provider available"));
-        }
-        if (shutdownStarted) {
-            return CompletableFuture.failedFuture(new RuntimeException("Shutdown has already started"));
-        }
-        try {
-            return CompletableFuture.supplyAsync(() -> {
-                        Provider provider = checkNotNull(selectedProvider.get(), "Selected provider must not be null.");
-                        BaseHttpClient client = networkService.getHttpClient(provider.getBaseUrl(), userAgent, provider.getTransportType());
-                        httpClient = Optional.of(client);
-                        long ts = System.currentTimeMillis();
-                        String param = provider.getApiPath();
-                        try {
-                            log.info("Request reference time from {}", client.getBaseUrl() + "/" + param);
-                            String json = client.get(param, Optional.of(new Pair<>("User-Agent", userAgent)));
-                            JsonNode timeNode = JsonMapperProvider.get().readTree(json).get("time");
-                            if (timeNode == null || timeNode.isNull()) {
-                                throw new RuntimeException("Response JSON missing 'time' field");
-                            }
-                            long time = timeNode.asLong() * 1000;
-                            log.info("Received reference time {} from {}/{} after {} ms", new Date(time), client.getBaseUrl(), param, System.currentTimeMillis() - ts);
-                            selectedProvider.set(selectNextProvider());
-                            shutdownHttpClient(client);
-                            return time;
-                        } catch (Exception e) {
-                            shutdownHttpClient(client);
-                            if (shutdownStarted) {
-                                throw new RuntimeException("Shutdown has already started");
-                            }
-
-                            Throwable rootCause = ExceptionUtil.getRootCause(e);
-                            log.warn("Encountered exception requesting reference time from provider {}", provider.getBaseUrl(), rootCause);
-
-                            if (rootCause instanceof HttpException httpException) {
-                                int responseCode = httpException.getResponseCode();
-                                // If not server error we pass the error to the client
-                                // 408 (Request Timeout) and 429 (Too Many Requests) are usually transient
-                                // and should rotate to another provider.
-                                if (responseCode < 500 && responseCode != 408 && responseCode != 429) {
-                                    throw new CompletionException(e);
-                                }
-                            }
-
-                            int numRecursions = recursionDepth.incrementAndGet();
-                            if (numRecursions < numTotalCandidates && failedProviders.size() < numTotalCandidates) {
-                                failedProviders.add(provider);
-                                selectedProvider.set(selectNextProvider());
-                                log.warn("We retry the request with new provider {}", selectedProvider.get().getBaseUrl());
-                                throw new RetryException("Retrying with next provider", recursionDepth);
-                            } else {
-                                log.warn("We exhausted all possible providers and give up");
-                                throw new RuntimeException("We failed at all possible providers and give up");
-                            }
-                        }
-                    }, POOL)
-                    .completeOnTimeout(null, conf.getTimeoutInSeconds(), SECONDS)
-                    .thenCompose(time -> {
-                        if (time == null) {
-                            return CompletableFuture.failedFuture(new RetryException("Timeout", recursionDepth));
-                        }
-                        return CompletableFuture.completedFuture(time);
-                    });
-        } catch (RejectedExecutionException e) {
-            log.error("Executor rejected requesting reference time task.", e);
-            return CompletableFuture.failedFuture(e);
-        }
-    }
-
-    private void shutdownHttpClient(BaseHttpClient client) {
-        try {
-            client.shutdown();
-        } catch (Exception ignore) {
-        }
+        return executeWithRetry(() -> makeRequest(provider -> {
+            long ts = System.currentTimeMillis();
+            String param = provider.getApiPath();
+            log.info("Request reference time from {}", provider.getBaseUrl() + "/" + param);
+            
+            BaseHttpClient client = networkService.getHttpClient(provider.getBaseUrl(), userAgent, provider.getTransportType());
+            try {
+                String json = client.get(param, Optional.of(new Pair<>("User-Agent", userAgent)));
+                JsonNode timeNode = JsonMapperProvider.get().readTree(json).get("time");
+                if (timeNode == null || timeNode.isNull()) {
+                    throw new RuntimeException("Response JSON missing 'time' field");
+                }
+                long time = timeNode.asLong() * 1000;
+                log.info("Received reference time {} from {}/{} after {} ms", new Date(time), provider.getBaseUrl(), param, System.currentTimeMillis() - ts);
+                
+                return CompletableFuture.completedFuture(time);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }));
     }
 }
